:ok_hand: Fixed zombie routines in checkpoint save on S3
